local_issues_lib

Struct Issues

pub struct Issues { /* private fields */ }

A collection of issues.

Implementations

impl Issues

pub fn new() -> Self

Creates a new Issues collection.

This initializes the collection with a โrootโ issue.

pub fn add_new_issue(&mut self, i: Issue) -> usize

Adds a new issue to the collection.

Returns the index of the newly added issue.

pub fn get_list(&self) -> &Vec<Issue>

Returns a reference to the list of issues.

pub fn find_from_title<T: AsRef<str>>( &mut self, s: T, ) -> Option<Vec<&mut Issue>>

Finds issues containing the given title string.

Returns Some(Vec<&mut Issue>) if matches are found, otherwise None.

pub fn find_from_updated_time( &mut self, st: DateTime<Local>, ed: DateTime<Local>, ) -> Option<Vec<&mut Issue>>

Finds issues updated within the given time range.

Returns Some(Vec<&mut Issue>) if matches are found, otherwise None.

pub fn find_from_created_time( &mut self, st: DateTime<Local>, ed: DateTime<Local>, ) -> Option<Vec<&mut Issue>>

Finds issues created within the given time range.

Returns Some(Vec<&mut Issue>) if matches are found, otherwise None.

pub fn find_from_comments<T: AsRef<str>>( &mut self, s: T, ) -> Option<Vec<&mut Issue>>

Finds issues containing a comment with the given text.

Returns Some(Vec<&mut Issue>) if matches are found, otherwise None.

pub fn get(&self, index: usize) -> Option<&Issue>

Gets an issue by its index.

pub fn get_mut(&mut self, index: usize) -> Option<&mut Issue>

Gets a mutable reference to an issue by its index.

pub fn fork(&mut self, from: usize) -> Option<usize>

Forks an issue from a given index.

The original issue is marked as CloseAsForked, and a new copy is created with from set to the original index.

Returns Some(usize) which is the index of the new forked issue, or None if the original issue doesnโt exist.

pub fn get_registered_labels(&self) -> Vec<String>

Returns a list of all registered labels.